+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## unreleased
+### Added
+- `TimeStringParser` - very handy to use in expansions so that you can write guard clauses like `[ t1After(3 min) ]`.
+
+---
+
 ## [0.8.4-alpha]
 ### Added
 - GilToCsharp - method pointers instead of function pointers so that expansions rarely needed.