+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.8.7-alpha]
 ### Added
 - Added support for automatically resolving state name conflicts.
   - Defaults to manual conflict resolution for now, but will be enabled by default in the future.